+++ pybop/costs/base_cost.py
@@ -21,48 +21,43 @@ class BaseCost:
     problem : object
         A problem instance containing the data and functions necessary for
         evaluating the cost function.
-    _target : array-like
+    target : array-like
         An array containing the target data to fit.
     n_outputs : int
         The number of outputs in the model.
-    _has_separable_problem : bool
+    has_separable_problem : bool
         If True, the problem is separable from the cost function and will be
         evaluated in advance of the call to self.compute() (default: False).
+    _de : float
+        The gradient of the cost function to use if an error occurs during
+        evaluation. Defaults to 1.0.
     """
 
     def __init__(self, problem: Optional[BaseProblem] = None):
-        self.parameters = Parameters()
-        self.transformation = None
+        self._parameters = Parameters()
+        self._transformation = None
         self.problem = problem
         self.verbose = False
         self._has_separable_problem = False
-        self.update_capacity = False
         self.y = None
         self.dy = None
         self._de = 1.0
         if isinstance(self.problem, BaseProblem):
             self._target = self.problem.target
-            self.parameters.join(self.problem.parameters)
+            self._parameters.join(self.problem.parameters)
             self.n_outputs = self.problem.n_outputs
             self.signal = self.problem.signal
-            self.transformation = self.parameters.construct_transformation()
+            self._transformation = self._parameters.construct_transformation()
             self._has_separable_problem = True
             self.grad_fail = None
             self.set_fail_gradient()
 
-    @property
-    def n_parameters(self):
-        return len(self.parameters)
-
-    @property
-    def has_separable_problem(self):
-        return self._has_separable_problem
-
-    @property
-    def target(self):
-        return self._target
-
-    def __call__(self, inputs: Union[Inputs, list], calculate_grad: bool = False):
+    def __call__(
+        self,
+        inputs: Union[Inputs, list],
+        calculate_grad: bool = False,
+        apply_transform: bool = False,
+    ):
         """
         This method calls the forward model via problem.evaluate(inputs),
         and computes the cost for the given output by calling self.compute().
@@ -71,6 +66,8 @@ class BaseCost:
         ----------
         inputs : Inputs or array-like
             The parameters for which to compute the cost and gradient.
+        calculate_grad : bool, optional
+            A bool condition designating whether to calculate the gradient.
 
         Returns
         -------
@@ -82,29 +79,27 @@ class BaseCost:
         ValueError
             If an error occurs during the calculation of the cost.
         """
-        if self.transformation:
-            p = self.transformation.to_model(inputs)
-        inputs = self.parameters.verify(p if self.transformation else inputs)
+        # Apply transformation if needed
+        # Note, we use the transformation and parameter properties here to enable
+        # differing attributes within the `LogPosterior` class
+        self.has_transform = self.transformation is not None and apply_transform
+        if self.has_transform:
+            inputs = self.transformation.to_model(inputs)
+        inputs = self.parameters.verify(inputs)
         self.parameters.update(values=list(inputs.values()))
+
         y, dy = None, None
+        if self._has_separable_problem:
+            if calculate_grad:
+                y, dy = self.problem.evaluateS1(self.problem.parameters.as_dict())
+                cost, grad = self.compute(y, dy=dy, calculate_grad=calculate_grad)
+                if self.has_transform and np.isfinite(cost):
+                    jac = self.transformation.jacobian(inputs)
+                    grad = np.matmul(grad, jac)
+                return cost, grad
 
-        try:
-            if self._has_separable_problem:
-                if calculate_grad is True:
-                    y, dy = self.problem.evaluateS1(self.problem.parameters.as_dict())
-                else:
-                    y = self.problem.evaluate(
-                        self.problem.parameters.as_dict(),
-                        update_capacity=self.update_capacity,
-                    )
-
-            return self.compute(y, dy=dy, calculate_grad=calculate_grad)
-
-        except NotImplementedError as e:
-            raise e
-
-        except Exception as e:
-            raise ValueError(f"Error in cost calculation: {e}") from e
+            y = self.problem.evaluate(self.problem.parameters.as_dict())
+        return self.compute(y, dy=dy, calculate_grad=calculate_grad)
 
     def compute(self, y: dict, dy: ndarray, calculate_grad: bool = False):
         """
@@ -178,3 +173,40 @@ class BaseCost:
             raise ValueError(
                 "Forward model sensitivities need to be provided alongside `calculate_grad=True` for `cost.compute`."
             )
+
+    def join_parameters(self, parameters):
+        """
+        Setter for joining parameters. This method sets the fail gradient if the join adds parameters.
+        """
+        original_n_params = self.n_parameters
+        self._parameters.join(parameters)
+        if original_n_params != self.n_parameters:
+            self.set_fail_gradient()
+
+    @property
+    def n_parameters(self):
+        return len(self._parameters)
+
+    @property
+    def has_separable_problem(self):
+        return self._has_separable_problem
+
+    @property
+    def target(self):
+        return self._target
+
+    @property
+    def parameters(self):
+        return self._parameters
+
+    @parameters.setter
+    def parameters(self, parameters):
+        self._parameters = parameters
+
+    @property
+    def transformation(self):
+        return self._transformation
+
+    @transformation.setter
+    def transformation(self, transformation):
+        self._transformation = transformation